Update msw to 1.0.X Fixes #8340
