+++ frontend-react/package.json
@@ -15,19 +15,16 @@
         "@tanstack/react-query-devtools": "^4.20.9",
         "@trussworks/react-uswds": "^4.0.0",
         "@uswds/uswds": "^3.1.0",
-        "@xmldom/xmldom": "0.8.6",
-        "axios": "^1.2.1",
+        "axios": "^1.3.3",
         "classnames": "^2.3.2",
         "dompurify": "^2.4.3",
         "downloadjs": "^1.4.7",
-        "loader-utils": "3.2.1",
         "lodash.chunk": "^4.2.0",
         "lodash.omit": "^4.5.0",
         "lodash.pick": "^4.4.0",
         "lodash.range": "^3.2.0",
         "lodash.uniq": "^4.5.0",
         "moment": "^2.29.4",
-        "otplib": "^12.0.1",
         "react": "<18.0.0",
         "react-dom": "<18.0.0",
         "react-helmet-async": "^1.3.0",
@@ -37,13 +34,11 @@
         "react-router-dom": "^6.3.0",
         "react-scroll-sync": "^0.11.0",
         "react-toastify": "^9.1.1",
-        "recursive-readdir": "2.2.3",
         "rehype-raw": "^6.1.1",
         "rehype-slug": "^5.0.1",
         "remark-gfm": "^3.0.1",
         "remark-toc": "^8.0.1",
         "rest-hooks": "^6.1.7",
-        "rimraf": "^3.0.2",
         "use-deep-compare-effect": "^1.8.1",
         "web-vitals": "^2.1.2"
     },
@@ -243,10 +238,13 @@
         "eslint-plugin-testing-library": "^5.9.1",
         "eslint-plugin-unused-imports": "2.0.0",
         "jest-canvas-mock": "^2.3.1",
-        "msw": "^0.36.8",
+        "msw": "^1.0.1",
         "npm-run-all": "^4.1.5",
+        "otplib": "^12.0.1",
         "prettier": "^2.8.3",
         "react-scripts": "^5.0.1",
+        "recursive-readdir": "2.2.3",
+        "rimraf": "^3.0.2",
         "sass": "^1.57.1",
         "typescript": "^4.9.4"
     },
