+++ frontend-react/src/components/Admin/AdminReceiverDashboard.tsx
@@ -547,11 +547,13 @@ function renderAllReceiverRows(props: {
                     </div>
                 </Grid>
                 <ScrollSyncPane enabled>
-                    <Grid row className={"horizontal-scroll"}>
-                        <Grid row className={"week-column"}>
-                            {perDayElements}
+                    <>
+                        <Grid row className={"horizontal-scroll"}>
+                            <Grid row className={"week-column"}>
+                                {perDayElements}
+                            </Grid>
                         </Grid>
-                    </Grid>
+                    </>
                 </ScrollSyncPane>
             </Grid>
         );
@@ -657,7 +659,6 @@ function MainRender(props: {
     }
 
     return (
-        //
         <ScrollSync horizontal enabled>
             <GridContainer className={"rs-admindash-component"}>
                 {FilterRenderedRows({
