+++ frontend-react/src/components/Admin/EditReceiverSettings.test.tsx
@@ -1,4 +1,5 @@
 import { fireEvent, screen } from "@testing-library/react";
+import { rest } from "msw";
 
 import { renderWithBase } from "../../utils/CustomRenderUtils";
 import { settingsServer } from "../../__mocks__/SettingsMockServer";
@@ -72,16 +73,23 @@ jest.mock("react-router-dom", () => ({
     },
     useParams: () => {
         return {
-            orgName: "abbott",
-            senderName: "user1234",
+            orgname: "abbott",
+            receivername: "user1234",
             action: "edit",
         };
     },
 }));
 
 describe("EditReceiverSettings", () => {
-    beforeAll(() => settingsServer.listen());
-    afterEach(() => settingsServer.resetHandlers());
+    beforeAll(() => {
+        settingsServer.listen();
+        settingsServer.use(
+            rest.get(
+                "https://test.prime.cdc.gov/api/settings/organizations/abbott/receivers/user1234",
+                (req, res, ctx) => res(ctx.json(mockData))
+            )
+        );
+    });
     afterAll(() => settingsServer.close());
     beforeEach(() => {
         renderWithBase(<EditReceiverSettings />);
