+++ frontend-react/src/components/Admin/EditSenderSettings.test.tsx
@@ -1,4 +1,5 @@
 import { fireEvent, screen } from "@testing-library/react";
+import { rest } from "msw";
 
 import { renderWithBase } from "../../utils/CustomRenderUtils";
 import OrgSenderSettingsResource from "../../resources/OrgSenderSettingsResource";
@@ -33,8 +34,8 @@ jest.mock("react-router-dom", () => ({
     },
     useParams: () => {
         return {
-            orgName: "abbott",
-            senderName: "user1234",
+            orgname: "abbott",
+            sendername: "user1234",
             action: "edit",
         };
     },
@@ -64,8 +65,15 @@ const testKeys = JSON.stringify([
 const testProcessingType = "sync";
 
 describe("EditSenderSettings", () => {
-    beforeAll(() => settingsServer.listen());
-    afterEach(() => settingsServer.resetHandlers());
+    beforeAll(() => {
+        settingsServer.listen();
+        settingsServer.use(
+            rest.get(
+                "https://test.prime.cdc.gov/api/settings/organizations/abbott/senders/user1234",
+                (req, res, ctx) => res(ctx.json(mockData))
+            )
+        );
+    });
     afterAll(() => settingsServer.close());
     beforeEach(() => {
         renderWithBase(<EditSenderSettings />);
