--- frontend-react/src/hooks/network/History/DeliveryHooks.test.ts
@@ -1,152 +1,0 @@
-import { renderHook } from "@testing-library/react-hooks";
-
-import { mockSessionContext } from "../../../contexts/__mocks__/SessionContext";
-import { MemberType } from "../../UseOktaMemberships";
-import { deliveryServer } from "../../../__mocks__/DeliveriesMockServer";
-import { QueryWrapper } from "../../../utils/CustomRenderUtils";
-import { Organizations } from "../../UseAdminSafeOrganizationName";
-
-import {
-    useReportsDetail,
-    useOrgDeliveries,
-    useReportsFacilities,
-} from "./DeliveryHooks";
-
-describe("DeliveryHooks", () => {
-    beforeAll(() => deliveryServer.listen());
-    afterEach(() => deliveryServer.resetHandlers());
-    afterAll(() => deliveryServer.close());
-    test("useReportsList", async () => {
-        mockSessionContext.mockReturnValue({
-            oktaToken: {
-                accessToken: "TOKEN",
-            },
-            activeMembership: {
-                memberType: MemberType.RECEIVER,
-                parsedName: "testOrg",
-            },
-            dispatch: () => {},
-            initialized: true,
-            isUserAdmin: false,
-            isUserReceiver: true,
-            isUserSender: false,
-        });
-
-        const { result } = renderHook(() => useOrgDeliveries("testService"));
-        const fetchResults = await result.current.fetchResults(" ", 10);
-        expect(fetchResults).toHaveLength(3);
-        expect(result.current.filterManager.pageSettings.currentPage).toEqual(
-            1
-        );
-        expect(result.current.filterManager.pageSettings.size).toEqual(10);
-        expect(result.current.filterManager.rangeSettings.from).toEqual(
-            "2000-01-01T00:00:00.000Z"
-        );
-        expect(result.current.filterManager.rangeSettings.to).toEqual(
-            "3000-01-01T00:00:00.000Z"
-        );
-        expect(result.current.filterManager.sortSettings.order).toEqual("DESC");
-    });
-    test("useReportDetail", async () => {
-        mockSessionContext.mockReturnValue({
-            oktaToken: {
-                accessToken: "TOKEN",
-            },
-            activeMembership: {
-                memberType: MemberType.RECEIVER,
-                parsedName: "testOrg",
-            },
-            dispatch: () => {},
-            initialized: true,
-            isUserAdmin: false,
-            isUserReceiver: true,
-            isUserSender: false,
-        });
-        const { result, waitForNextUpdate } = renderHook(
-            () => useReportsDetail("123"),
-            { wrapper: QueryWrapper() }
-        );
-        await waitForNextUpdate();
-        expect(result.current.reportDetail?.reportId).toEqual("123");
-    });
-    test("useReportFacilities", async () => {
-        mockSessionContext.mockReturnValue({
-            oktaToken: {
-                accessToken: "TOKEN",
-            },
-            activeMembership: {
-                memberType: MemberType.RECEIVER,
-                parsedName: "testOrg",
-            },
-            dispatch: () => {},
-            initialized: true,
-            isUserAdmin: false,
-            isUserReceiver: true,
-            isUserSender: false,
-        });
-        const { result, waitForNextUpdate } = renderHook(
-            () => useReportsFacilities("123"),
-            { wrapper: QueryWrapper() }
-        );
-        await waitForNextUpdate();
-        expect(result.current.reportFacilities?.length).toEqual(2);
-    });
-
-    describe("useOrgDeliveries", () => {
-        describe("when requesting as a receiver", () => {
-            beforeEach(() => {
-                mockSessionContext.mockReturnValue({
-                    oktaToken: {
-                        accessToken: "TOKEN",
-                    },
-                    activeMembership: {
-                        memberType: MemberType.RECEIVER,
-                        parsedName: "testOrg",
-                    },
-                    dispatch: () => {},
-                    initialized: true,
-                    isUserAdmin: false,
-                    isUserReceiver: true,
-                    isUserSender: false,
-                });
-            });
-
-            test("fetchResults returns an array of deliveries", async () => {
-                const { result } = renderHook(() =>
-                    useOrgDeliveries("testService")
-                );
-                const results = await result.current.fetchResults(" ", 10);
-
-                expect(results).toHaveLength(3);
-            });
-        });
-
-        describe("when requesting as an admin", () => {
-            beforeEach(() => {
-                mockSessionContext.mockReturnValue({
-                    oktaToken: {
-                        accessToken: "TOKEN",
-                    },
-                    activeMembership: {
-                        memberType: MemberType.PRIME_ADMIN,
-                        parsedName: Organizations.PRIMEADMINS,
-                    },
-                    dispatch: () => {},
-                    initialized: true,
-                    isUserAdmin: true,
-                    isUserReceiver: false,
-                    isUserSender: false,
-                });
-            });
-
-            test("fetchResults returns an empty array", async () => {
-                const { result } = renderHook(() =>
-                    useOrgDeliveries("testService")
-                );
-                const results = await result.current.fetchResults(" ", 10);
-
-                expect(results).toHaveLength(0);
-            });
-        });
-    });
-});
