+++ frontend-react/src/pages/Validate.test.tsx
@@ -1,14 +1,14 @@
 import React, { Suspense } from "react";
 
-import { renderWithSession } from "../utils/CustomRenderUtils";
 import Spinner from "../components/Spinner";
+import { renderWithFullAppContext } from "../utils/CustomRenderUtils";
 
 import Validate from "./Validate";
 
 describe("Validate", () => {
     test("Renders with no errors", async () => {
         expect(() =>
-            renderWithSession(
+            renderWithFullAppContext(
                 <Suspense fallback={<Spinner size={"fullpage"} />}>
                     <Validate />
                 </Suspense>
